Front: Generic Infinite List component
